+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Documentation site available at https://davnavr.github.io/FSharpWrap/
+### Fixed
+- Add check to avoid generating members whose names end with "$W" on computation expression types
 
 ## [0.4.0] - 2020-11-24
 ### Added